+++ Core/Executor/ContentTypeManager.php
@@ -448,16 +448,15 @@ class ContentTypeManager extends RepositoryExecutor implements MigrationGenerato
     }
 
     /**
-     * @param string $matchType
-     * @param string|string[] $matchValue
+     * @param array $matchCondition
      * @param string $mode
      * @throws \Exception
      * @return array
      */
-    public function generateMigration($matchType, $matchValue, $mode)
+    public function generateMigration(array $matchCondition, $mode)
     {
         $previousUserId = $this->loginUser(self::ADMIN_USER_ID);
-        $contentTypeCollection = $this->contentTypeMatcher->match(array($matchType => $matchValue));
+        $contentTypeCollection = $this->contentTypeMatcher->match($matchCondition);
         $data = array();
 
         /** @var \eZ\Publish\API\Repository\Values\ContentType\ContentType $contentType */