+++ Core/Executor/RoleManager.php
@@ -316,16 +316,15 @@ class RoleManager extends RepositoryExecutor implements MigrationGeneratorInterf
     }
 
     /**
-     * @param string $matchType
-     * @param string|string[] $matchValue
+     * @param array $matchCondition
      * @param string $mode
      * @throws \Exception
      * @return array
      */
-    public function generateMigration($matchType, $matchValue, $mode)
+    public function generateMigration(array $matchCondition, $mode)
     {
         $previousUserId = $this->loginUser(self::ADMIN_USER_ID);
-        $roleCollection = $this->roleMatcher->match(array($matchType => $matchValue));
+        $roleCollection = $this->roleMatcher->match($matchCondition);
         $data = array();
 
         /** @var \eZ\Publish\API\Repository\Values\User\Role $role */