+++ Core/Matcher/ObjectStateMatcher.php
@@ -15,7 +15,7 @@ class ObjectStateMatcher extends RepositoryMatcher implements KeyMatcherInterfac
     const MATCH_OBJECTSTATE_IDENTIFIER = 'objectstate_identifier';
 
     protected $allowedConditions = array(
-        self::MATCH_ALL,
+        self::MATCH_ALL, self::MATCH_NOT,
         self::MATCH_OBJECTSTATE_ID, self::MATCH_OBJECTSTATE_IDENTIFIER,
         // aliases
         'id', 'identifier'
@@ -56,6 +56,9 @@ class ObjectStateMatcher extends RepositoryMatcher implements KeyMatcherInterfac
 
                 case self::MATCH_ALL:
                     return new ObjectStateCollection($this->findAllObjectStates());
+
+                case self::MATCH_NOT:
+                    return new ObjectStateCollection(array_diff_key($this->findAllObjectStates(), $this->matchObjectState($values)->getArrayCopy()));
             }
         }
     }