+++ Core/Matcher/SectionMatcher.php
@@ -14,7 +14,7 @@ class SectionMatcher extends RepositoryMatcher implements KeyMatcherInterface
     const MATCH_SECTION_IDENTIFIER = 'section_identifier';
 
     protected $allowedConditions = array(
-        self::MATCH_ALL,
+        self::MATCH_ALL, self::MATCH_NOT,
         self::MATCH_SECTION_ID, self::MATCH_SECTION_IDENTIFIER,
         // aliases
         'id', 'identifier'
@@ -55,6 +55,9 @@ class SectionMatcher extends RepositoryMatcher implements KeyMatcherInterface
 
                 case self::MATCH_ALL:
                     return new SectionCollection($this->findAllSections());
+
+                case self::MATCH_NOT:
+                    return new SectionCollection(array_diff_key($this->findAllSections(), $this->matchSection($values)->getArrayCopy()));
             }
         }
     }